fixat settings view
